+++ nebula/core/situationalawareness/awareness/sanetwork/neighborpolicies/starneighborpolicy.py
@@ -1,4 +1,4 @@
-from nebula.core.situationalawareness.awareness.neighborpolicies.neighborpolicy import NeighborPolicy
+from nebula.core.situationalawareness.awareness.sanetwork.neighborpolicies.neighborpolicy import NeighborPolicy
 from nebula.core.utils.locker import Locker
 
 class STARNeighborPolicy(NeighborPolicy):
